requirements.txt
python_gamelocker==0.5.0
crayons==0.1.2
requests==2.18.2
typing==3.6.2
